services/ingestion/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
nats-py==2.6.0
python-json-logger==2.0.7
